% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_params_set_local.R
\name{board_params_set_local}
\alias{board_params_set_local}
\title{Generate Properly Formatted Board Parameters}
\usage{
board_params_set_local(board_alias, url, folder = "")
}
\arguments{
\item{board_alias}{name of the board}

\item{url}{url of the local directory}

\item{folder}{path of the study folder where the data product will be stored}
}
\value{
A data.frame board_params
}
\description{
This function simply provides a consistent format for the board
parameters
}
\examples{
\dontrun{
board_params_set_labkey(board_alias = "xxxx",
  folder = "xxxx")
}
}
