% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPI.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Produce a symmetric correlation matrix from values.}
\usage{
cor_matrix(...)
}
\arguments{
\item{...}{Correlation values to transform into the symmetric correlation matrix (by row).}
}
\value{
Return a symmetric correlation matrix.
}
\description{
Produce a symmetric correlation matrix from values.
}
\examples{
cor_matrix(
  1.0, 0.7, 0.3,
  0.7, 1.0, 0.5,
  0.3, 0.5, 1.0
)

cor_matrix(
  1.0, NA, NA,
  0.7, 1.0, NA,
  0.3, 0.5, 1.0
)

}
