% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds_set_aws.R
\name{creds_set_aws}
\alias{creds_set_aws}
\title{Create formatted credentials for connecting to an s3 bucket board}
\usage{
creds_set_aws(
  profile_name = character(0),
  key = character(0),
  secret = character(0)
)
}
\arguments{
\item{profile_name}{The name of an AWS credentials profile for the s3 bucket
where the data product is stored. This is a name associated with a set of
credentials within the .aws/credentials files in your home directory. See
the \href{https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-files.html}{AWS CLI credentials file documentation}
for more information.}

\item{key}{The AWS Access Key ID associated with your access to the s3 bucket
where the data product is stored.}

\item{secret}{The AWS Secret Access Key associated with your access to the s3
bucket where the data product is stored.}
}
\value{
A data.frame with class "aws_creds" and a column for each param.
}
\description{
Build a data frame that contains all of the credentials
needed to connect to an s3 bucket pin board. Additionally, it ensures the
minimum requirement that either a named profile or a set of AWS key and
secret are provided. If a named profile is provided, the key and secret
will be ignored.
}
\examples{
\dontrun{
aws_creds <- creds_set_aws(
  key = Sys.getenv("AWS_KEY"),
  secret = Sys.getenv("AWS_SECRET")
)
}
}
