% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds_set_local.R
\name{creds_set_local}
\alias{creds_set_local}
\title{Generate Local Credentials}
\usage{
creds_set_local(cache)
}
\arguments{
\item{api_key}{Local API key for the user}
}
\value{
local_creds
}
\description{
Generate Local Credentials
}
\examples{
\dontrun{creds_set_local(api_key = Sys.getenv("LABKEY_API_KEY"))}
}
