% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPI.R
\name{data_random}
\alias{data_random}
\title{Generate random data.}
\usage{
data_random(k, n, seed = NULL)
}
\arguments{
\item{k}{Number of variables.}

\item{n}{Number of observations (cases).}

\item{seed}{Random seed for replicable results.
Defaults to \code{NULL}.}
}
\value{
Return a data.frame of random data.
}
\description{
Generate random data.
}
\examples{
d = data_random(k=5, n=100, seed=1)
cor_network(d)

}
