% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_get.R
\name{dp_get}
\alias{dp_get}
\title{Get the Data Object}
\usage{
dp_get(board_params, data_name, version = NULL)
}
\arguments{
\item{board_params}{use \code{board_params_set_s3} or \code{board_params_set_labkey}
for this. It contains the parameters for the board on which the data product
is pinned}

\item{data_name}{data name on the board}

\item{version}{data version pinned, default gets you the latest. Executed
dp_list to see what's available}
}
\value{
dp pinned data
}
\description{
Load into working environment the data data object
}
\examples{
\dontrun{
aws_creds <- creds_set_aws(key = Sys.getenv("AWS_KEY"),
secret = Sys.getenv("AWS_SECRET"))
board_params <- board_params_set_s3(board_alias = "board_alias",
  bucket_name = "bucket_name", region = "us-east-1")
dp_connect(board_params, aws_creds)
dp <- dp_get("dp-study-branch", board_params = board_params)
}
}
