% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_make_params.R
\name{dp_make_params}
\alias{dp_make_params}
\title{Make dp params to connect to the Data Product Board}
\usage{
dp_make_params(url, repo_token)
}
\arguments{
\item{url}{Repository URL}

\item{repo_token}{Personal Access Token (PAT) needed to access the contents of
the private repository}
}
\value{
Named list of dp params
}
\description{
Make dp params to connect to the Data Product Board. This is needed
prior to interacting with the content of the board.
}
