% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dpconnect_check}
\alias{dpconnect_check}
\title{Check dpconnect executed}
\usage{
dpconnect_check(board_params)
}
\arguments{
\item{board_params}{board_params (only the alias needed)}
}
\description{
This checks state whether dpconnect is already executed
}
\keyword{internal}
