% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpi-package.R
\docType{package}
\name{dpi-package}
\alias{dpi}
\alias{dpi-package}
\title{dpi: Data Product Programming Interface}
\description{
This package provides an interface to data product.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/amashadihossein/dpi/issues}
}

}
\author{
\strong{Maintainer}: Afshin Mashadi-Hossein \email{amashadihossein@gmail.com}

Authors:
\itemize{
  \item Yue Jiang \email{rivehill@gmail.com}
}

}
\keyword{internal}
