% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPI.R
\name{matrix_cor}
\alias{matrix_cor}
\title{Produce a symmetric correlation matrix from values.}
\usage{
matrix_cor(...)
}
\arguments{
\item{...}{Correlation values to transform into the symmetric correlation matrix (by row).}
}
\value{
Return a symmetric correlation matrix.
}
\description{
Produce a symmetric correlation matrix from values.
}
\examples{
matrix_cor(
  1.0, 0.7, 0.3,
  0.7, 1.0, 0.5,
  0.3, 0.5, 1.0
)

}
