% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPI.R
\name{plot.dpi.curve}
\alias{plot.dpi.curve}
\title{[S3 method] Plot DPI curve analysis results.}
\usage{
\method{plot}{dpi.curve}(x, file = NULL, width = 6, height = 4, dpi = 500, ...)
}
\arguments{
\item{x}{A data.frame (of new class \code{dpi.curve})
returned from \code{\link{DPI_curve}}.}

\item{file}{File name of saved plot (\code{".png"} or \code{".pdf"}).}

\item{width, height}{Width and height (in inches) of saved plot.
Defaults to \code{6} and \code{4}.}

\item{dpi}{Dots per inch (figure resolution). Defaults to \code{500}.}

\item{...}{Other arguments (currently not used).}
}
\value{
Return a \code{ggplot} object.
}
\description{
[S3 method] Plot DPI curve analysis results.
}
