% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPI.R
\name{plot.dpi}
\alias{plot.dpi}
\title{[S3 method] Plot DPI results.}
\usage{
\method{plot}{dpi}(x, ...)
}
\arguments{
\item{x}{A data.frame (of new class \code{dpi})
returned from \code{\link{DPI}}.}

\item{...}{Other arguments (currently not used).}
}
\value{
Return a \code{ggplot} object.
}
\description{
[S3 method] Plot DPI results.
}
