% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPI.R
\name{print.dpi}
\alias{print.dpi}
\title{[S3 method] Print DPI summary and plot.}
\usage{
\method{print}{dpi}(x, digits = 3, ...)
}
\arguments{
\item{x}{A data.frame (of new class \code{dpi})
returned from \code{\link{DPI}}.}

\item{digits}{Number of decimal places. Defaults to \code{3}.}

\item{...}{Other arguments (currently not used).}
}
\value{
No return value.
}
\description{
[S3 method] Print DPI summary and plot.
}
