% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPI.R
\name{random_data}
\alias{random_data}
\title{Generate random data.}
\usage{
random_data(n.var, n.obs, seed = NULL)
}
\arguments{
\item{n.var}{Number of variables.}

\item{n.obs}{Number of observations (cases).}

\item{seed}{Random seed for replicable results.
Defaults to \code{NULL}.}
}
\value{
Return a data.frame of random data.
}
\description{
Generate random data.
}
