% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPI.R
\name{summary.dpi}
\alias{summary.dpi}
\title{[S3 method] Summarize DPI results.}
\usage{
\method{summary}{dpi}(object, ...)
}
\arguments{
\item{object}{A data.frame (of new class \code{dpi})
returned from \code{\link{DPI}}.}

\item{...}{Other arguments (currently not used).}
}
\value{
Return a list with new class \code{summary.dpi} of
summarized results and raw DPI data.frame.
}
\description{
[S3 method] Summarize DPI results.
}
