% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_params_set_labkey.R
\name{board_params_set_labkey}
\alias{board_params_set_labkey}
\title{Create formatted parameters that specify a LabKey board}
\usage{
board_params_set_labkey(
  board_alias = deprecated(),
  url,
  folder = "",
  cache_alias = NULL
)
}
\arguments{
\item{board_alias}{A short name for the board.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} this argument is deprecated with
pins ≥ 1.0.}

\item{url}{The URL of the LabKey server.}

\item{folder}{The path of the study folder on the LabKey server where the
data product will be stored.}

\item{cache_alias}{A short name for the cache folder (optional).}
}
\value{
A data.frame with class "labkey_board" and a column for each param.
}
\description{
Build a data frame that contains all of the parameters
needed to specify a LabKey pin board.
}
\examples{
\dontrun{
board_params_set_labkey(
  url = "https://learn.labkey.com/",
  folder = "folder_name"
)
}
}
