% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_params_set_labkey.R
\name{board_params_set_labkey}
\alias{board_params_set_labkey}
\title{Generate properly formatted board parameters for Labkey board}
\usage{
board_params_set_labkey(board_alias, url, folder = "")
}
\arguments{
\item{board_alias}{name of the board}

\item{url}{url of the labkey server}

\item{folder}{path of the study folder where the data product will be stored}
}
\value{
A data.frame board_params
}
\description{
This function simply provides a consistent format for the board
parameters
}
\examples{
\dontrun{
board_params_set_labkey(
  board_alias = "xxxx",
  folder = "xxxx"
)
}
}
