% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_params_set_labkey.R
\name{board_params_set_labkey}
\alias{board_params_set_labkey}
\title{Create formatted parameters that specify a LabKey board}
\usage{
board_params_set_labkey(board_alias, url, folder = "")
}
\arguments{
\item{board_alias}{A short name for the board.}

\item{url}{The URL of the LabKey server.}

\item{folder}{The path of the study folder on the LabKey server where the
data product will be stored.}
}
\value{
A data.frame with class "labkey_board" and a column for each param.
}
\description{
Build a data frame that contains all of the parameters
needed to specify a LabKey pin board.
}
\examples{
\dontrun{
board_params_set_labkey(
  board_alias = "labkey-board",
  url = "https://learn.labkey.com/",
  folder = "folder_name"
)
}
}
