% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_params_set_local.R
\name{board_params_set_local}
\alias{board_params_set_local}
\title{Generate properly formatted board parameters for local board}
\usage{
board_params_set_local(folder, board_alias = deprecated())
}
\arguments{
\item{folder}{path to data product storage folder. NOTE: as a general
guidance this folder should be treated as a permanent location with immutable
content. For example, a data product project folder \emph{is not} an appropriate
directory.}

\item{board_alias}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} this argument is deprecated with newer pins}
}
\value{
A data.frame with properly formatted board_params
}
\description{
This function simply provides a consistent format for the board
parameters
}
\examples{
\dontrun{
board_params_set_local(folder = "xxxx")
}
}
