% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_params_set_local.R
\name{board_params_set_local}
\alias{board_params_set_local}
\title{Create formatted parameters that specify a local storage board}
\usage{
board_params_set_local(folder, board_alias = deprecated())
}
\arguments{
\item{folder}{The path to the local storage folder where the data product
will be stored. NOTE: this folder should be treated as a permanent location
with immutable content. For example, a data product git repo project folder
\emph{is not} an appropriate directory.}

\item{board_alias}{A short name for the board.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} this argument is deprecated with
pins ≥ 1.0.}
}
\value{
A data.frame with class "local_board" and a column for each param.
}
\description{
Build a data frame that contains all of the parameters
needed to specify a local storage pin board.
}
\examples{
\dontrun{
board_params_set_local(folder = "xxxx")
}
}
