% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_params_set_s3.R
\name{board_params_set_s3}
\alias{board_params_set_s3}
\title{Create formatted parameters that specify an s3 bucket board}
\usage{
board_params_set_s3(bucket_name, region, board_alias = deprecated())
}
\arguments{
\item{bucket_name}{The name of the s3 bucket where the data product will be
stored.}

\item{region}{The AWS region where the s3 bucket was created, e.g.
"us-east-1" or "us-west-1".}

\item{board_alias}{A short name for the board.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} this argument is deprecated with
pins ≥ 1.0.}
}
\value{
A data.frame with class "s3_board" and a column for each param.
}
\description{
Build a data frame that contains all of the parameters
needed to specify an s3 bucket pin board.
}
\examples{
\dontrun{
board_params_set_s3(
  bucket_name = "bucket_name",
  region = "us-east-1"
)
}
}
