% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_params_set_s3.R
\name{board_params_set_s3}
\alias{board_params_set_s3}
\title{Generate Properly Formatted Board Parameters}
\usage{
board_params_set_s3(bucket_name, region, board_alias = deprecated())
}
\arguments{
\item{bucket_name}{name of the s3 bucket}

\item{region}{AWS region for the s3 bucket e.g. "us-east-1" or "us-west-1"}

\item{board_alias}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} this argument is deprecated with newer pins}
}
\value{
A data.frame with properly formatted board_params
}
\description{
This function simply provides a consistent format for the board
parameters
}
\examples{
\dontrun{
board_params_set_s3(
  bucket_name = "bucket_name",
  region = "us-east-1"
)
}
}
