% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_params_set_s3.R
\name{board_params_set_s3}
\alias{board_params_set_s3}
\title{Generate Properly Formatted Board Parameters}
\usage{
board_params_set_s3(board_alias, bucket_name, region)
}
\arguments{
\item{board_alias}{name of the board}

\item{bucket_name}{name of the s3 bucket}

\item{region}{AWS region for the s3 bucket e.g. "us-east-1" or "us-west-1"}
}
\value{
A data.frame board_params
}
\description{
This function simply provides a consistent format for the board
parameters
}
\examples{
\dontrun{
board_params_set_s3(
  board_alias = "board_alias",
  bucket_name = "bucket_name", region = "us-east-1"
)
}
}
