% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds_set_aws.R
\name{creds_set_aws}
\alias{creds_set_aws}
\title{Generate Properly Formatted AWS Credentials}
\usage{
creds_set_aws(
  profile_name = character(0),
  key = character(0),
  secret = character(0)
)
}
\arguments{
\item{profile_name}{AWS named profile for the bucket housing the data
product. This is a name associated with credential within .aws/credentials
(See AWS CLI configuring named profiles).}

\item{key}{aws key associated with your access to the bucket housing the
data product}

\item{secret}{aws secret associated with your access to the bucket housing
the data product}
}
\value{
A data.frame aws_creds containing properly formatted AWS credentials
}
\description{
This function simply provides a consistent format for aws
credentials. Additionally, it ensures the minimum requirement of either
named profile or a set of AWS (key, secret) are provided. Note if named
profile is provided key and secret will be ignored.
}
\examples{
\dontrun{
aws_creds <- creds_set_aws(key = Sys.getenv("AWS_KEY"),
secret = Sys.getenv("AWS_SECRET"))
}
}
