% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds_set_labkey.R
\name{creds_set_labkey}
\alias{creds_set_labkey}
\title{Generate LabKey Credentials}
\usage{
creds_set_labkey(api_key)
}
\arguments{
\item{api_key}{LabKey API key for the user}
}
\value{
labkey_creds
}
\description{
Get credentials for LabKey board
}
\examples{
\dontrun{
creds_set_labkey(api_key = Sys.getenv("LABKEY_API_KEY"))
}
}
