% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds_set_labkey.R
\name{creds_set_labkey}
\alias{creds_set_labkey}
\title{Generate properly formatted Labkey credentials}
\usage{
creds_set_labkey(api_key)
}
\arguments{
\item{api_key}{Labkey API key for the user}
}
\value{
labkey_creds
}
\description{
Generate properly formatted Labkey credentials
}
\examples{
\dontrun{creds_set_labkey (api_key = Sys.getenv("LABKEY_API_KEY"))}
}
