% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_connect.R
\name{dp_connect}
\alias{dp_connect}
\title{Connect to the data product board}
\usage{
dp_connect(board_params, creds, ...)
}
\arguments{
\item{board_params}{use \code{board_params_set_s3} or \code{board_params_set_labkey}
for this. It contains the parameters for the board on which the data product
is pinned}

\item{creds}{When \code{local_board}, creds is ignored and need to be specified.
Otherwise, use creds_set_aws() or creds_set_labkey() to set this. It should
contain credentials needed to access the remote where the dp is pinned.}

\item{...}{other parameters}
}
\value{
TRUE
}
\description{
Connect to the board housing the data product. This is needed
prior to interacting with the content of the board.
}
\examples{
\dontrun{
aws_creds <- creds_set_aws(key = Sys.getenv("AWS_KEY"),
secret = Sys.getenv("AWS_SECRET"))
board_params <- board_params_set_s3(board_alias = "board_alias",
  bucket_name = "bucket_name", region = "us-east-1")
dp_connect(board_params, aws_creds)
}
}
