% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_connect.R
\name{dp_connect}
\alias{dp_connect}
\title{Connect to the Data Product Board}
\usage{
dp_connect(board_params, creds, ...)
}
\arguments{
\item{board_params}{use \code{board_params_set_s3}, \code{board_params_set_labkey}, or
\code{board_params_set_local} to specify board parameters. It contains the information
for the board on which the data product is pinned.}

\item{creds}{use \code{creds_set_aws} or \code{creds_set_labkey} to set this. When using
a local board, creds is ignored and does not need to be specified.}

\item{...}{other parameters}
}
\value{
data product object
}
\description{
Connect to the board housing the data product. This is needed
prior to interacting with the content of the board.
}
\examples{
\dontrun{
aws_creds <- creds_set_aws(
  key = Sys.getenv("AWS_KEY"),
  secret = Sys.getenv("AWS_SECRET")
)
board_params <- board_params_set_s3(
  bucket_name = "bucket_name",
  region = "us-east-1"
)
board_object <- dp_connect(board_params, aws_creds)
}
}
