% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_get.R
\name{dp_get}
\alias{dp_get}
\title{Get the data product from the remote pin board}
\usage{
dp_get(board_object, data_name, version = NULL)
}
\arguments{
\item{board_object}{A \code{pins_board} object from \code{dp_connect}.}

\item{data_name}{The name of the data product to get from the remote pin
board, i.e. "dp-cars-us001". To get a list of data products available on
the remote pin board, use \code{dp_list}.}

\item{version}{The hash specifying the data product version to get from the
remote pin board. If \code{NULL}, \code{dp_get} will get the latest version
available. To get a list of versions available for each data product on the
remote pin board, use \code{dp_list}.}
}
\value{
A data product object, which is a list with class \code{dp} and items
\code{input}, \code{output}, and \code{README}.
}
\description{
Get a data product object from the provided remote pin board,
including input links and output data structure. By default get the latest
version, or get a particular version by specifying the version hash (find
available versions using \code{dp_list}).
}
\examples{
\dontrun{
aws_creds <- creds_set_aws(
  key = Sys.getenv("AWS_KEY"),
  secret = Sys.getenv("AWS_SECRET")
)
board_params <- board_params_set_s3(
  bucket_name = "bucket_name",
  region = "us-east-1"
)
board_object <- dp_connect(board_params, aws_creds)
dp <- dp_get(board_object, data_name = "data-name")
}
}
