% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_get.R
\name{dp_get}
\alias{dp_get}
\title{Get the Data Object}
\usage{
dp_get(board_object, data_name, version = NULL)
}
\arguments{
\item{board_object}{board object from \code{dp_connect}}

\item{data_name}{name of the data product on the board, i.e. dp-cars-us001. To
get a list of available data products, use \code{dp_list}}

\item{version}{data version to retrieve. If not specified, will retrieve the latest
version. Use \code{dp_list} to see available versions.}
}
\value{
data product object
}
\description{
Load into working environment the data product object
}
\examples{
\dontrun{
aws_creds <- creds_set_aws(
  key = Sys.getenv("AWS_KEY"),
  secret = Sys.getenv("AWS_SECRET")
)
board_params <- board_params_set_s3(
  bucket_name = "bucket_name",
  region = "us-east-1"
)
board_object <- dp_connect(board_params, aws_creds)
dp <- dp_get(board_object, "dp-study-branch", data_name = "data-name")
}
}
