% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_list.R
\name{dp_list}
\alias{dp_list}
\title{List Data Products on a Board}
\usage{
dp_list(board_object)
}
\arguments{
\item{board_object}{board object from \code{dp_connect}}
}
\value{
a tibble containing metadata of all versions of data products on board
}
\description{
List all data products on a named board you are connected with.
It requires connection via \code{dp_connect} first.
}
\examples{
\dontrun{
aws_creds <- creds_set_aws(
  key = Sys.getenv("AWS_KEY"),
  secret = Sys.getenv("AWS_SECRET")
)
board_params <- board_params_set_s3(
  board_alias = "board_alias",
  bucket_name = "bucket_name",
  region = "us-east-1"
)
board_object <- dp_connect(board_params, aws_creds)
dp_list(board_object)
}

}
