% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_list.R
\name{dp_list}
\alias{dp_list}
\title{List data products on the board}
\usage{
dp_list(board_params)
}
\arguments{
\item{board_params}{use \code{board_params_set_s3} or \code{board_params_set_labkey}
for this. It contains the parameters for the board on which the data product
is pinned}
}
\value{
all data products on the board
}
\description{
List all data products on a named board you are connected with.
It requires connection via dp_connet first.
}
\examples{
\dontrun{
aws_creds <- creds_set_aws(
  key = Sys.getenv("AWS_KEY"),
  secret = Sys.getenv("AWS_SECRET")
)
board_params <- board_params_set_s3(
  board_alias = "board_alias",
  bucket_name = "bucket_name", region = "us-east-1"
)
dp_connect(board_params, aws_creds)
dp_list(board_params)
}

}
