% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPI.R
\name{DPI}
\alias{DPI}
\title{The Directed Prediction Index (DPI).}
\usage{
DPI(
  model,
  y,
  x,
  data = NULL,
  n.cov = 0,
  n.sample = 1000,
  seed = NULL,
  progress
)
}
\arguments{
\item{model}{Model object (\code{lm}).}

\item{y}{Dependent (outcome) variable.}

\item{x}{Independent (predictor) variable.}

\item{data}{[Optional] Defaults to \code{NULL}.
If \code{data} is specified, then \code{model} will be ignored and
a linear model \code{lm({y} ~ {x} + .)} will be fitted inside.
This is helpful for exploring all variables in a dataset.}

\item{n.cov}{Number of random covariates
(simulating potential omitted variables)
added to each simulation sample.
\itemize{
\item Defaults to \code{0}: in such case, bootstrap samples
(resampling with replacement) are used for simulation.
\item If \code{n.cov > 0}, then the raw data (without bootstrapping)
are used, with \code{n.cov} random variables appended,
for simulation.
\item Test different \code{n.cov} values as robustness checks.
}}

\item{n.sample}{Number of simulation samples.
Defaults to \code{1000}.}

\item{seed}{Random seed for replicable results.
Defaults to \code{NULL}.}

\item{progress}{Show progress bar.
Defaults to \code{FALSE} (if \code{n.sample < 5000}).}
}
\value{
Return a data.frame of simulation results:
\itemize{
\item \code{DPI}
\itemize{
\item \code{t.beta.xy^2 * (R2.Y - R2.X)}
}
\item \code{t.beta.xy}
\itemize{
\item \emph{t} value for coefficient of X predicting Y (always equal to \emph{t} value for coefficient of Y predicting X) when controlling for all other covariates
}
\item \code{df.beta.xy}
\itemize{
\item residual degree of freedom (df) of \code{t.beta.xy}
}
\item \code{r.partial.xy}
\itemize{
\item partial correlation (always with the same \emph{t} value as \code{t.beta.xy}) between X and Y when controlling for all other covariates
}
\item \code{delta.R2}
\itemize{
\item \code{R2.Y - R2.X}
}
\item \code{R2.Y}
\itemize{
\item \eqn{R^2} of regression model predicting Y using X and all other covariates
}
\item \code{R2.X}
\itemize{
\item \eqn{R^2} of regression model predicting X using Y and all other covariates
}
}
}
\description{
The Directed Prediction Index (DPI) is a simulation-based method
for quantifying \emph{endogeneity} of outcome vs. predictor variables
in linear regression models.
}
\details{
\eqn{
\begin{aligned}
\text{DPI}_{X \rightarrow Y} & = t^2 \cdot \Delta R^2 \\
& = t_{\beta_{XY|Covs}}^2 \cdot (R_{Y \sim X + Covs}^2 - R_{X \sim Y + Covs}^2)
\end{aligned}
}

Since \eqn{t_{\beta} = t_{partial.r}} and when all variables are standardized:
\eqn{
\begin{aligned}
\text{DPI}_{X \rightarrow Y} & = t_{partial.r_{XY|Covs}}^2 \cdot \left( \frac{\sum_{i=1}^n (\hat{Y}_i - \bar{Y})^2}{\sum_{i=1}^n (Y_i - \bar{Y})^2} - \frac{\sum_{i=1}^n (\hat{X}_i - \bar{X})^2}{\sum_{i=1}^n (X_i - \bar{X})^2} \right) \\
& = \frac{pr_{XY|Covs}^2 \cdot (n - k_{Covs} - 2)}{1 - pr_{XY|Covs}^2} \cdot \left( \frac{\sum_{i=1}^n \hat{Y}_i^2}{\sum_{i=1}^n Y_i^2} - \frac{\sum_{i=1}^n \hat{X}_i^2}{\sum_{i=1}^n X_i^2} \right)
\end{aligned}
}
}
\examples{
\donttest{model = lm(Temp ~ ., data=airquality)
DPI(model, y="Temp", x="Solar.R")  # bootstrap sample if n.cov=0
DPI(model, y="Temp", x="Solar.R", n.cov=1, seed=1)  # raw sample

DPI(data=airquality, y="Temp", x="Solar.R", n.cov=10, seed=1)
}
}
